Feature/frontend login register configuration

+++ frontend/src/pages/RegisterForm.jsx
@@ -1,21 +1,29 @@
-import { useState } from "react";
-import { Link } from 'react-router-dom';
-import { loginUser } from "../services/userService";
+import { useState, useContext } from "react";
+import { Link, useNavigate } from 'react-router-dom';
+import { loginUser, registerUser } from "../services/userService";
+import { AuthContext } from "../contexts/AuthContext";
 
-function RegisterForm({ setUser }) {
+function RegisterForm() {
+
+  const { setUser } = useContext(AuthContext);
 
   const [isRegister, setIsRegister] = useState(false)
 
   const [email, setEmail] = useState("")
   const [password, setPassword] = useState("")
+  const [confirmPassword, setConfirmPassword] = useState("")
 
   const [firstName, setFirstName] = useState("")
   const [lastName, setLastName] = useState("")
-  const [role, setRole] = useState("GUARDIAN");
+  const [role, setRole] = useState("GUARDIAN")
 
-  const [error, setError] = useState(null);
+  const [error, setError] = useState(null)
+
+  const navigate = useNavigate()
 
   function handleChange(input) {
+    setError(null);
+
     const { name, value } = input.target
     switch (name) {
       case "email":
@@ -23,6 +31,9 @@ function RegisterForm({ setUser }) {
         break;
       case "password":
         setPassword(value);
+        break;
+      case "confirmPassword":
+        setConfirmPassword(value);
         break;
       case "firstName":
         setFirstName(value);
@@ -38,21 +49,44 @@ function RegisterForm({ setUser }) {
     }
   }
 
-  function handleSubmit(form) {
-    
+  async function handleSubmit(form) {
     form.preventDefault()
+
     if (isRegister) {
-      console.log("Registrando:", { email, password, firstName, lastName });
-      setUser({ username: firstName, role, loggedIn: true });
-      navigate("/map");
+      if (confirmPassword !== password) {
+        setError("Las contrasenas no coinciden.")
+
+      } else {
+        try {
+          const data = await registerUser(firstName, lastName, email, password, role);
+          console.log("Token from server:", data.token);
+          localStorage.setItem("token", data.token);
+          localStorage.setItem("user", JSON.stringify({
+            username: data.firstName,
+            role: data.role
+          }));
+          setUser({ username: data.firstName, role: data.role, loggedIn: true });
+          navigate("/profile");
+
+        } catch (err) {
+          setError(err.message)
+        }
+      }
 
     } else {
 
       try {
-        const data = loginUser(email, password);
-        localStorage.setItem("token", data.token);      // guardar JWT
-        setUser({ username: data.username, role: data.role, loggedIn: true });
+        const data = await loginUser(email, password);
+        localStorage.setItem("token", data.token);
+        localStorage.setItem("user", JSON.stringify({
+          username: data.firstName,
+          role: data.role
+        }));
+        console.log(data);
+
+        setUser({ username: data.firstName, role: data.role, loggedIn: true });
         navigate("/map");
+
       } catch (err) {
         setError(err.message);
       }
@@ -63,7 +97,7 @@ function RegisterForm({ setUser }) {
 
     <div>
 
-      <h1>{isRegister ? "REGISTRO" : "LOGIN"}</h1>
+      <h1>{isRegister ? "REGISTRO" : "LOG IN"}</h1>
 
       <form onSubmit={handleSubmit}>
         {isRegister && (
@@ -123,10 +157,10 @@ function RegisterForm({ setUser }) {
 
         {isRegister && (
           <input
-            type="confirm-password"
+            type="password"
             placeholder="Confirmar contrasena"
-            name="password"
-            value={password}
+            name="confirmPassword"
+            value={confirmPassword}
             onChange={handleChange}
             required
           />)}
@@ -136,10 +170,16 @@ function RegisterForm({ setUser }) {
         <button type="submit">{isRegister ? "REGÍSTRATE" : "LOGIN"}</button>
       </form>
 
+      {error && (
+        <div className="error-message">
+          {error}
+        </div>
+      )}
+
       <br />
 
       <p>
-        {isRegister ? "¿Ya tienes cuenta?" : "¿No tienes cuenta?"}{" "}
+        {isRegister ? "¿Ya tienes cuenta?" : "¿No tienes cuenta?"}
         <button onClick={() => setIsRegister(!isRegister)}>
           {isRegister ? "INICIA SESIÓN" : "REGÍSTRATE"}
         </button>
